🔄 Refactor, Fix, and Name Change in API: From /api/session to /api/user. Added 'create register with passport' method in services, controller, and repository.
