+++ src/views/login.handlebars
@@ -2,7 +2,7 @@
   <h1>Inicio de Sesion</h1>
 
   <div>
-    <form action="/api/session/login" method="post">
+    <form action="/api/user/login" method="post">
       <label for="email">Correo</label><br />
       <input type="email" name="email" id="email" placeholder="introduzca su correo electronico" />
       <br />
@@ -18,6 +18,6 @@
 
   <div>
     <br>
-    <a href="/api/session/github"> <button type="submit">Entrar con GITHUB</button></a>
+    <a href="/api/user/github"> <button type="submit">Entrar con GITHUB</button></a>
   </div>
 </div>
