+++ src/views/products.handlebars
@@ -19,7 +19,7 @@
 <p>Visit : {{this.visit}}</p>
 <div>
   <h3>eliminando la Cuenta</h3>
-  <form action="/api/session/logout" method="get">
+  <form action="/api/user/logout" method="get">
     <button type="submit">clearSession</button>
   </form>
