+++ src/views/profile.handlebars
@@ -6,7 +6,7 @@
   <p>{{this.lastname}}</p>
 
   <div>
-    <form action="/api/session/logout" method="get">
+    <form action="/api/user/logout" method="get">
       <br />
       <button type="submit">logout</button>
     </form>
