+++ src/views/recover.handlebars
@@ -2,7 +2,7 @@
   <h1>Recuperacion de Contraseña</h1>
 
   <div>
-    <form action="/api/session/recover-psw" method="post">
+    <form action="/api/user/recover-psw" method="post">
     <label for="email">Correo</label><br>
     <input 
       type="email"
