+++ src/views/register.handlebars
@@ -2,7 +2,7 @@
   <h1>REGISTRO</h1>
 
   <div>
-    <form action="/api/session/register" method="post">
+    <form action="/api/user/register" method="post">
       <label for="nombre-cliente">Nombre del usuario</label><br />
       <input
         type="text"
